rename prompt_schema.py to prompt.py
