+++ README.md
@@ -70,7 +70,7 @@ The `Context` class is used to pass in the context for the LLM prompt. A `Contex
 
 To run `uniflow` with the default instructions and few-shot examples, you can pass in a list of `Context` objects to the flow. For example:
 ```
-from uniflow.op.prompt_schema import Context
+from uniflow.op.prompt import Context
 
 data = [
     Context(
@@ -104,7 +104,7 @@ Once you've decided on your `Config` and prompting strategy, you can run the flo
     ```
     from uniflow.flow.client import TransformClient
     from uniflow.flow.config import TransformOpenAIConfig, OpenAIModelConfig
-    from uniflow.op.prompt_schema import Context
+    from uniflow.op.prompt import Context
     ```
 1. Preprocess your data in to chunks to pass into the flow. In the future we will have `Preprocessing` flows to help with this step, but for now you can use a library of your choice, like [pypdf](https://pypi.org/project/pypdf/), to chunk your data.
     ```
@@ -121,7 +121,7 @@ Once you've decided on your `Config` and prompting strategy, you can run the flo
 
 1. [Optional] If you want to use a customized instruction and/or examples, create a `PromptTemplate`.
     ```
-    from uniflow.op.prompt_schema import PromptTemplate
+    from uniflow.op.prompt import PromptTemplate
 
     guided_prompt = PromptTemplate(
     instruction="Generate a one sentence summary based on the last context below. Follow the format of the examples below to include context and summary in the response",
@@ -213,7 +213,7 @@ Here is an example of how to pass in a custom configuration to the `Client` obje
 ```
 from uniflow.flow.client import TransformClient
 from uniflow.flow.config import TransformOpenAIConfig, OpenAIModelConfig
-from uniflow.op.prompt_schema import Context
+from uniflow.op.prompt import Context
 
 
 contexts = ["It was a sunny day and the sky color is blue.", "My name is bobby and I am a talent software engineer working on AI/ML."]
