+++ uniflow/flow/transform/transform_openai_flow.py
@@ -6,7 +6,7 @@ from uniflow.flow.flow import Flow
 from uniflow.node import Node
 from uniflow.op.model.llm_processor import JsonFormattedDataProcessor, LLMDataProcessor
 from uniflow.op.model.model_op import ModelOp
-from uniflow.op.prompt_schema import PromptTemplate
+from uniflow.op.prompt import PromptTemplate
 
 
 class OpenAIModelFlow(Flow):
